updated to pull strings from internal pdf structure
