+++ plugins/pdf.py
@@ -1,4 +1,6 @@
 import sys, os.path, re
+import StringIO
+import xml.etree.ElementTree as ET
 from pdfminer.psparser import PSKeyword, PSLiteral, LIT
 from pdfminer.pdfparser import PDFParser
 from pdfminer.pdfdocument import PDFDocument, PDFNoOutlines
@@ -6,11 +8,12 @@ from pdfminer.pdftypes import PDFObjectNotFound, PDFValueError
 from pdfminer.pdftypes import PDFStream, PDFObjRef, resolve1, stream_value
 from pdfminer.pdfpage import PDFPage
 from pdfminer.utils import isnumber
-from pdfid import PDFiD2String, PDFiD
+from pdfid import PDFiD2String, PDFiD #Source code for pdfid put in public domain by Didier Stevens, no Copyright
 
 
 
 class Plugin:
+
 
     __NAME__ = 'pdf'
 
@@ -19,24 +22,35 @@ class Plugin:
         self.analyzed_mimes = ['application/pdf']
 
 
+    def analyze(self, afile):
+        '''Analyze PDF files and extract metadata about the file into the 
+        FileAnalysis object.
 
-    SC_PAT = re.compile(r'[\000-\037&<>()"\042\047\134\177-\377]')
-    def e(self,s):
-        return ESC_PAT.sub(lambda m:'&#%d;' % ord(m.group(0)), s)
+        Args:
+            afile (FileAnalysis): The file to be analyzed.
+        
+        Returns:
+            None
+        '''
+
+        #Adapted from pdfminer utility dumppdf.py
+        ESC_PAT = re.compile(r'[\000-\037&<>()"\042\047\134\177-\377]')
+        def e(s):
+            return ESC_PAT.sub(lambda m:'&#%d;' % ord(m.group(0)), s)
 
 
-    # dumpxml
-    def dumpxml(self,out, obj, codec=None):
-        if obj is None:
-            out.write('<null />')
-            return
+        # dumpxml
+        def dumpxml(out, obj, codec=None):
+            if obj is None:
+                out.write('<null />')
+                return
 
             if isinstance(obj, dict):
                 out.write('<dict size="%d">\n' % len(obj))
                 for (k,v) in obj.iteritems():
                     out.write('<key>%s</key>\n' % k)
                     out.write('<value>')
-                    self.dumpxml(out, v)
+                    dumpxml(out, v)
                     out.write('</value>\n')
                 out.write('</dict>')
                 return
@@ -44,13 +58,13 @@ class Plugin:
             if isinstance(obj, list):
                 out.write('<list size="%d">\n' % len(obj))
                 for v in obj:
-                    self.dumpxml(out, v)
+                    dumpxml(out, v)
                     out.write('\n')
                 out.write('</list>')
                 return
 
             if isinstance(obj, str):
-                out.write('<string size="%d">%s</string>' % (len(obj), self.e(obj)))
+                out.write('<string size="%d">%s</string>' % (len(obj), e(obj)))
                 return
 
             if isinstance(obj, PDFStream):
@@ -60,11 +74,11 @@ class Plugin:
                     out.write(obj.get_data())
                 else:
                     out.write('<stream>\n<props>\n')
-                    self.dumpxml(out, obj.attrs)
+                    dumpxml(out, obj.attrs)
                     out.write('\n</props>\n')
                     if codec == 'text':
                         data = obj.get_data()
-                        out.write('<data size="%d">%s</data>\n' % (len(data), self.e(data)))
+                        out.write('<data size="%d">%s</data>\n' % (len(data), e(data)))
                     out.write('</stream>')
                 return
 
@@ -86,65 +100,54 @@ class Plugin:
 
             raise TypeError(obj)
 
-    # dumptrailers
-    def dumptrailers(self,out, doc):
-        for xref in doc.xrefs:
-            out.write('<trailer>\n')
-            self.dumpxml(out, xref.trailer)
-            out.write('\n</trailer>\n\n')
-        return
+        # dumptrailers
+        def dumptrailers(out, doc):
+            for xref in doc.xrefs:
+                out.write('<trailer>\n')
+                dumpxml(out, xref.trailer)
+                out.write('\n</trailer>\n\n')
+            return
 
-    # dumpallobjs
-    def dumpallobjs(self,out, doc, codec=None):
-        visited = set()
-        out.write('<pdf>')
-        for xref in doc.xrefs:
-            for objid in xref.get_objids():
-                if objid in visited: continue
-                visited.add(objid)
-                try:
-                    obj = doc.getobj(objid)
-                    if obj is None: continue
-                    out.write('<object id="%d">\n' % objid)
-                    self.dumpxml(out, obj, codec=codec)
-                    out.write('\n</object>\n\n')
-                except PDFObjectNotFound as e:
-                    print >>sys.stderr, 'not found: %r' % e
-        self.dumptrailers(out, doc)
-        out.write('</pdf>')
-        return
-
-
-    def pdf_id(self,afile):
-        result = PDFiD2String(PDFiD(afile.path),True)
-        # Split off of new lines
-        lines = result.split('\n')[1:]
-        for line in lines:
-            #strip white spaces
-            line = line.strip()
-            #parse out line into key,value by sequential white spaces
-            kv_pair = re.split('\s+',line)
-            if len(kv_pair) > 1:
-                #remove forward slash
-                key = re.sub('/','',kv_pair[0])
-                value = kv_pair[1]
-                # if we have more than 2 entries then the value was parsed incorrectly, join the other entries in list into one value
-                if len(kv_pair) > 2:
-                    value = ' '.join(kv_pair[1:])
-                    # set the attribute with our key,value
-                setattr(afile,key,value)
+        # dumpallobjs
+        def dumpallobjs(out, doc, codec=None):
+            visited = set()
+            out.write('<pdf>')
+            for xref in doc.xrefs:
+                for objid in xref.get_objids():
+                    if objid in visited: continue
+                    visited.add(objid)
+                    try:
+                        obj = doc.getobj(objid)
+                        if obj is None: continue
+                        out.write('<object id="%d">\n' % objid)
+                        dumpxml(out, obj, codec=codec)
+                        out.write('\n</object>\n\n')
+                    except PDFObjectNotFound as e:
+                        print >>sys.stderr, 'not found: %r' % e
+            dumptrailers(out, doc)
+            out.write('</pdf>')
+            return out.getvalue()
 
 
-    def analyze(self, afile):
-        '''Analyze PDF files and extract metadata about the file into the 
-        FileAnalysis object.
+        def pdf_id(afile):
+            result = PDFiD2String(PDFiD(afile.path),True)
+            # Split off of new lines
+            lines = result.split('\n')[1:]
+            for line in lines:
+                #strip white spaces
+                line = line.strip()
+                #parse out line into key,value by sequential white spaces
+                kv_pair = re.split('\s+',line)
+                if len(kv_pair) > 1:
+                    #remove forward slash
+                    key = re.sub('/','',kv_pair[0])
+                    value = kv_pair[1]
+                    # if we have more than 2 entries then the value was parsed incorrectly, join the other entries in list into one value
+                    if len(kv_pair) > 2:
+                        value = ' '.join(kv_pair[1:])
+                        # set the attribute with our key,value
+                    setattr(afile,key,value)
 
-        Args:
-            afile (FileAnalysis): The file to be analyzed.
-        
-        Returns:
-            None
-        '''
         if afile.mime_type in self.analyzed_mimes:
 
             # Parse the metadata for the pdf file and add all pdf metadata
@@ -161,12 +164,25 @@ class Plugin:
                 process_metadata = False
 
             if process_metadata:
+
+                pdf_tags = ['JavaScript','OpenAction']
+
                 # count out PDF tags
-                self.pdf_id(afile)
+                pdf_id(afile)
 
                 parser = PDFParser(fp)
                 doc = PDFDocument(parser,'')
+                xml_io = StringIO.StringIO()
+                xml_str = dumpallobjs(xml_io,doc,None)
+                xml_io.close()
 
-                self.dumpallobjs(sys.stdout,doc,None)
+                xml = ET.fromstring(xml_str)
+
+                strings = list()
+                for child in xml.iter('string'):
+                        if child.text not in strings:
+                            strings.append(child.text)
+
+                setattr(afile,'strings',strings)
